[SPARK-40432][SS][PYTHON] Introduce GroupStateImpl and GroupStateTimeout in PySpark
### What changes were proposed in this pull request?

This PR introduces GroupStateImpl and GroupStateTimeout in PySpark, and updates Scala codebase to support convenient conversion between PySpark implementation and Scala implementation.

Co-authored with HyukjinKwon .

This is a breakdown PR of #37863.

### Why are the changes needed?

This change will be leveraged in SPARK-40434.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

N/A. We will make sure test suites are constructed via E2E manner under SPARK-40431.

Closes #37889 from HeartSaVioR/SPARK-40432.

Lead-authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Co-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>